embeddings/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.1+cpu
sentence-transformers==3.0.1
pandas==2.2.2
pyarrow==16.1.0
deltalake==0.18.1
s3fs==2024.6.0
faiss-cpu==1.9.0
numpy<2.0.0
python-dotenv==1.0.1
boto3==1.34.131